+++ src/dipdup/config.py
@@ -45,7 +45,7 @@ from dipdup.datasources.subscription import Subscription
 from dipdup.datasources.subscription import TransactionSubscription
 from dipdup.enums import OperationType
 from dipdup.enums import ReindexingAction
-from dipdup.enums import ReindexingReasonC
+from dipdup.enums import ReindexingReason
 from dipdup.enums import SkipHistory
 from dipdup.exceptions import ConfigInitializationException
 from dipdup.exceptions import ConfigurationError
@@ -958,10 +958,8 @@ default_hooks = {
 
 @dataclass
 class AdvancedConfig:
-    reindex: Dict[ReindexingReasonC, ReindexingAction] = field(default_factory=dict)
+    reindex: Dict[ReindexingReason, ReindexingAction] = field(default_factory=dict)
     scheduler: Optional[Dict[str, Any]] = None
-    # TODO: Drop in major version
-    oneshot: bool = False
     postpone_jobs: bool = False
     early_realtime: bool = False
     merge_subscriptions: bool = False