+++ src/dipdup/dipdup.py
@@ -203,7 +203,7 @@ class IndexDispatcher:
                     await index_state.save()
                 elif new_hash != index_state.config_hash:
                     await self._ctx.reindex(
-                        ReindexingReason.CONFIG_HASH_MISMATCH,
+                        ReindexingReason.config_modified,
                         old_hash=index_state.config_hash,
                         new_hash=new_hash,
                     )
@@ -212,7 +212,7 @@ class IndexDispatcher:
             elif template:
                 if template not in self._ctx.config.templates:
                     await self._ctx.reindex(
-                        ReindexingReason.MISSING_INDEX_TEMPLATE,
+                        ReindexingReason.config_modified,
                         index_name=index_state.name,
                         template=template,
                     )
@@ -401,14 +401,9 @@ class DipDup:
         # TODO: Fix Tortoise ORM to raise more specific exception
         except KeyError:
             try:
-                # TODO: Drop with major version bump
-                # NOTE: A small migration, ReindexingReason became ReversedEnum in 3.1.0
-                for item in ReindexingReason:
-                    await conn.execute_script(f'UPDATE dipdup_schema SET reindex = "{item.name}" WHERE reindex = "{item.value}"')
-
                 self._schema = await Schema.get_or_none(name=schema_name)
             except KeyError:
-                await self._ctx.reindex(ReindexingReason.SCHEMA_HASH_MISMATCH)
+                await self._ctx.reindex(ReindexingReason.schema_modified)
 
         # NOTE: Call even if Schema is present; there may be new tables
         await generate_schema(conn, schema_name)
@@ -424,14 +419,14 @@ class DipDup:
             try:
                 await self._schema.save()
             except OperationalError:
-                await self._ctx.reindex(ReindexingReason.SCHEMA_HASH_MISMATCH)
+                await self._ctx.reindex(ReindexingReason.schema_modified)
 
         elif not self._schema.hash:
             self._schema.hash = schema_hash  # type: ignore
             await self._schema.save()
 
         elif self._schema.hash != schema_hash:
-            await self._ctx.reindex(ReindexingReason.SCHEMA_HASH_MISMATCH)
+            await self._ctx.reindex(ReindexingReason.schema_modified)
 
         elif self._schema.reindex:
             await self._ctx.reindex(self._schema.reindex)
@@ -495,7 +490,7 @@ class DipDup:
             actual_head = await datasource.get_block(db_head.level)
             if db_head.hash != actual_head.hash:
                 await self._ctx.reindex(
-                    ReindexingReason.BLOCK_HASH_MISMATCH,
+                    ReindexingReason.rollback,
                     hash=db_head.hash,
                     actual_hash=actual_head.hash,
                 )